refactor: Move FakeIdentityGetter to testutils
